requirements.txt
# ── Core data ──────────────────────────────────────────────────────────────────
pandas>=2.0.0,<3.0.0
numpy>=1.24.0,<2.0.0
yfinance>=0.2.36
requests>=2.31.0
pytz>=2024.1
feedparser>=6.0.0

# ── NLP / Sentiment ────────────────────────────────────────────────────────────
transformers==4.53.2
tokenizers==0.21.2
torch>=2.2.0
sentencepiece>=0.1.99
tf-keras

# ── ML ─────────────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
scipy>=1.12.0

# ── API ────────────────────────────────────────────────────────────────────────
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
pydantic>=2.6.0

# ── Dashboard ──────────────────────────────────────────────────────────────────
streamlit>=1.33.0,<2.0.0
plotly>=5.20.0

# ── Dev ────────────────────────────────────────────────────────────────────────
python-dotenv>=1.0.0
tqdm>=4.66.0
